requirements.txt
streamlit>=1.33
pandas>=2.0
transformers>=4.41
torch>=2.1.0
sentencepiece>=0.1.99
python-Levenshtein>=0.21.1
huggingface_hub>=0.23
safetensors>=0.4.2
protobuf<5
